requirements.txt
flask
gspread
google-auth
python-dotenv
psycopg2
redis
rq
